Better created property and new summary-for property for daily summary capture.
Created property always reflects the time the entry was created,
regardless of whether the entry was invoked with a C-1 prefix to select
a date.

Added a summary-for property that contains the date for which the
summary was written, so you can more easily create sparse trees etc
based on date criteria.